backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langgraph>=0.2.0
langchain>=0.1.13
langchain-openai>=0.1.1
langchain-community>=0.0.29
vaderSentiment==3.3.2
openai>=1.14.2
google-generativeai>=0.8.0
chromadb>=0.5.0
redis>=5.0.0
playwright>=1.40.0
pandas>=2.2.1
python-dotenv==1.0.1
pydantic>=2.6.4
websockets>=12.0
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt==4.1.2
